pyproject.toml

[tool.poetry]
name = "dblp-service"
version = "0.1.0"
description = ""
authors = ["Adam C. Saunders <saunders@cs.umass.edu>"]
readme = "readme.md"
packages = [{include = "dblp_service"},]

[tool.poetry.dependencies]
python = "^3.11"
sparqlwrapper = "^2.0.0"
pytest = "^7.4.2"
pyright = "^1.1.331"
click = "^8.1.7"
rich = "^13.6.0"
bigtree = "^0.13.1"
pylint = "^3.0.1"
flake8 = "^6.1.0"
mypy = "^1.6.0"
black = "^23.9.1"
requests-futures = "^1.0.1"
bibtexparser = "^2.0.0b3"
marshmallow = "^3.20.1"
email-validator = "^2.0.0.post2"
flake8-pyproject = "^1.2.3"
icecream = "^2.1.3"
snoop = "^0.4.3"
birdseye = "^0.9.5"
types-requests = "^2.31.0.10"
pytest-asyncio = "^0.21.1"
tqdm = "^4.66.1"
types-tqdm = "^4.66.0.4"
beautifulsoup4 = "^4.12.2"
marshmallow-dataclass = "^8.6.0"
types-beautifulsoup4 = "^4.12.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

[tool.pyright]
useLibraryCodeForTypes = true
typeCheckingMode = "strict" # strict/none
exclude = [
    "submodules/**"
]


### https://github.com/microsoft/pyright/blob/main/docs/configuration.md#diagnostic-rule-defaults
reportMissingImports = "error" #[boolean or string, optional]
reportMissingTypeStubs = false #[boolean or string, optional]
reportUnusedImport = "warning" #[boolean or string, optional]
reportUnusedVariable = "information" #[boolean or string, optional]

# reportUntypedNamedTuple = false
# reportGeneralTypeIssues = "error" #[boolean or string, optional]
# reportPropertyTypeMismatch = "warning" #[boolean or string, optional]
# reportUntypedFunctionDecorator = false #[boolean or string, optional]
# reportUnknownParameterType = "warning" #[boolean or string, optional]
# reportUnknownArgumentType = "warning" #[boolean or string, optional]
# reportUnknownLambdaType = "warning" #[boolean or string, optional]
# reportUnknownVariableType = "warning" #[boolean or string, optional]
# reportUnknownMemberType = "none" #[boolean or string, optional]
# reportMissingParameterType = "warning" #[boolean or string, optional]
# reportMissingTypeArgument = "error" #[boolean or string, optional]
# reportMissingModuleSource = "none"
